fix bugs b/c of changing money data type from Double to BigDecimal + update database to version 10

+++ app/src/main/java/com/ssmmhh/jibam/models/TransactionsRecyclerViewItem.kt
@@ -22,8 +22,8 @@ sealed class TransactionsRecyclerViewItem(
 
     data class Header(
         val date: String,
-        val expensesSum: Double?,
-        val incomeSum: Double?
+        val expensesSum: BigDecimal?,
+        val incomeSum: BigDecimal?
     ) : TransactionsRecyclerViewItem(HEADER_VIEW_TYPE)
